+++ CHANGELOG.md
@@ -90,6 +90,8 @@
   - Specifying the combined services from HTTPRoutes feature via
     by `spec.translation.combinedServicesFromDifferentHTTPRoutes`
     [#1934](https://github.com/Kong/kong-operator/pull/1934)
+  - Specifying the drain support by `spec.translation.drainSupport`
+    [#1940](https://github.com/Kong/kong-operator/pull/1940)
 - Introduce flags `--apiserver-host` for API, `--apiserver-qps` and
   `--apiserver-burst` to control the QPS and burst (rate-limiting) for the
   Kubernetes API server client.